Cleanup some of the warnings

+++ acorn/Logger.cpp
@@ -479,7 +479,7 @@ void acorn::Logger::end_error(ErrCode error_code) {
         for (wchar_t wc : file.path) {
             narrow_path += static_cast<char>(wc);
         }
-        error_code_interceptor(error_code, narrow_path, start_line_number);
+        error_code_interceptor(error_code, narrow_path, static_cast<int>(start_line_number));
         mtx.unlock();
         return;
     }
